Add EdDSA signing/verification test
